+++ orthanc_api_client/__init__.py
@@ -3,6 +3,7 @@ from .exceptions import *
 from .helpers import *
 from .change import ChangeType, ResourceType
 from .study import Study, StudyInfo
+from .series import Series, SeriesInfo
 from .job import Job, JobInfo, JobType, JobStatus
 from .http_client import HttpClient
 from .downloaded_instance import DownloadedInstance