+++ orthanc_api_client/resources/__init__.py
@@ -1,5 +1,5 @@
 from .resources import Resources
 from .instances import Instances
-from .series import Series
+from .series_list import SeriesList
 from .studies import Studies
 from .jobs import Jobs