+++ release-notes.md
@@ -1,4 +1,8 @@
+v 0.9.0
+=======
+
 - **BREAKING CHANGE:** renamed `download_study` and `download_series` into `download_instances`
+- introduced `Series`, `SeriesInfo`, `Instance` and `InstanceInfo` classes
 
 v 0.8.3
 =======